P2 sist part funciona

+++ skeleton/UniformParticleGenerator.cpp
@@ -23,8 +23,8 @@ list<Particle*> UniformParticleGenerator::generateParticles()
 	list<Particle*>particles;
 	for (int i = 0; i < numParticles; i++) {
 		if (abs(d(generator)) < generationProb) {
-			Vector3 desvPos = { d(generator) * posWidth.x,d(generator) * posWidth.y ,d(generator) * posWidth.z };
-			Vector3 desvVel = { d(generator) * velWidth.x,d(generator) * velWidth.y ,d(generator) * velWidth.z };
+			Vector3 desvPos = { (float)d(generator) * posWidth.x,(float)d(generator) * posWidth.y ,(float)d(generator) * posWidth.z };
+			Vector3 desvVel = { (float)d(generator) * velWidth.x,(float)d(generator) * velWidth.y ,(float)d(generator) * velWidth.z };
 			Particle* part = new Particle(meanPos + desvPos, meanVel + desvVel, acc, damp, color);
 			part->setMass(mass);
 			part->setSpawnTime(GetLastTime());
